Cotizador_Coandes-copia/requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
Pillow
google-genai
google-generativeai